Fix build warning on macOS
The warning is `ld: warning: directory not found for option
'-L/usr/local/Cellar/libuv/1.11.0/lib'`

It is because the library is installed via homebrew and the homebrew
formula of libuv was upgraded to 1.12.0, and the library is installed
in `/usr/local/Cellar/libuv/1.12.0/`.

However, there is no need to include the path in build config because
symbolic link to the files are added to `/usr/local/lib`, which is
already included.